removed the powergraph reference implementations from the generated documentation.  it was confusing to have multiple implementations of the algorithms using different apis in the docs.  the mpgraph implementations are the only ones for which documentation is currently generated.
It would be nice to group the mpgraph code by namespaces or some similar mechanisms, even a marker interface, that makes it easier to identify the mpgraph code versus the bc40 code.

Another possibility is to not generate the bc40 documentation since it is below the level of the mpgraph api. However, some classes, such as CsrProblem, are used by user code and thus need to appear in the docs.