Adding the Apache 2 license file to the b40c directory.  This license file was not present in the b40c distribution.  Note that we actually have b40c under both BSD and Apache 2 and it might be desirable to include both licenses in this distribution.
Created a NOTICE.TXT file for b40c. There was none in the original distribution.  This notice file includes the license header and a notification that we have extensively modified the b40c code.

Adding the BSD license to the modern GPU directory. This license file was not present in the modern GPU code distribution.  I have created the file from the license text on the modern GPU web site.