I have added a "test" target to the Makefile that runs a variety of ground truth tests and random starting vertex tests (w/o validation).
All algorithms now terminate (exit(1)) if they do not pass validation.  They would exit with status code (0) before.  The new validation test suite relies on this new behavior.

I have added conditional compilation for gethostname(). It is not available on Ubuntu.  I have no idea why.

I have modified the largePerformanceGraphs directory to NOT apply the transform that was used for graphlab.  This is still available, but it is commented out in ufl.mk.  This means that we can now download the graphs and use them for self-validation (except that SSSP validation is too expensive).  I have also modified this to download the bitcoin.mtx file from dropbox.

I have fixed the CPU CC algorithm.

I have modified the CPU SSSP algorithm to support undirected graphs, but the implementation needs to be verified and it is ridiculously slow so we basically need to throw it out and write a work efficient SSSP implementation for the CPU that can handle both directed and undirected traversal. (Really, it should handle forward, reverse, and undirected).

All algorithms now support "-p stats=1" to report the vertex degree histogram.

I have reverted the code in enactor_vertex_centrix that was setting the mgpu device using cuCtxGetDevice(), The code might be valid, but I was having problems around this issue with the Java code.  I will look at this again soon.  The Algorithm Makefiles now include "-lcuda" which is necessary if we want to use the cuCtxGetDevice() call.

      {
//          CUdevice dev;
//          if(cuCtxGetDevice(&dev)) throw;
//          m_mgpuContext = mgpu::CreateCudaDevice(dev);
          const int dev = cfg.getParameter<int>("device");
          std::cerr<<"mgpuContext: device="<<dev<<std::endl;
          m_mgpuContext = mgpu::CreateCudaDevice(dev); // TODO Should avoid access to "device" parameter from inside MapGraph core.
      }

I have replaced omp_get_wtime() with time() and difftime().

The CsrGraph degree histogram now reports the maximum degree and a vertex having that maximum degree.

The matrix market reader was incorrectly rejecting files that had only a row and column and no edge weight. The code now silently uses an edge weight of one when none was given in the file.

I have verified MapGraph execution using "make test" for this commit.